# Swagger Docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Pweb-Backend-API",
        "version": "1.0",
        "title": "Swagger API",
        "contact": {
            "name": "Facultad Informática"
        }
    },
    "host": "localhost:8085",
    "basePath": "/",
...
    "paths": {
...
        "/api/v1/casillas/pisos/{num_piso}": {
            "get": {
                "tags": [
                    "casilla-controller"
                ],
                "summary": "getCasillasByPisoId",
                "operationId": "getCasillasByPisoIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "num_piso",
                        "in": "path",
                        "description": "num_piso",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/CasillaDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/v1/estantes/almacenes/{idalmacen}": {
            "get": {
                "tags": [
                    "estante-controller"
                ],
                "summary": "getEsantesByAlmacenId",
                "operationId": "getEsantesByAlmacenIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "idalmacen",
                        "in": "path",
                        "description": "idalmacen",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/EstanteDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/v1/estantes/{cod_estante}": {
            "get": {
                "tags": [
                    "estante-controller"
                ],
                "summary": "getEstanteById",
                "operationId": "getEstanteByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "cod_estante",
                        "in": "path",
                        "description": "cod_estante",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/EstanteDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/v1/pisos": {
            "get": {
                "tags": [
                    "piso-controller"
                ],
                "summary": "getPisos",
                "operationId": "getPisosUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PisoDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "piso-controller"
                ],
                "summary": "create",
                "operationId": "createUsingPOST_7",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "piso",
                        "description": "piso",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PisoDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "piso-controller"
                ],
                "summary": "update",
                "operationId": "updateUsingPUT_7",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "piso",
                        "description": "piso",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PisoDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/pisos/estantes/{cod_estante}": {
            "get": {
                "tags": [
                    "piso-controller"
                ],
                "summary": "getEsantesByAlmacenId",
                "operationId": "getEsantesByAlmacenIdUsingGET_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "cod_estante",
                        "in": "path",
                        "description": "cod_estante",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/PisoDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/pisos/{cod_almacen}/{cod_empresa}/{cod_estante}/{/numpiso}": {
            "delete": {
                "tags": [
                    "piso-controller"
                ],
                "summary": "delete",
                "operationId": "deleteUsingDELETE_7",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "cod_almacen",
                        "in": "path",
                        "description": "cod_almacen",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cod_empresa",
                        "in": "path",
                        "description": "cod_empresa",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "cod_estante",
                        "in": "path",
                        "description": "cod_estante",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "numpiso",
                        "in": "path",
                        "description": "numpiso",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/pisos/{numpiso}": {
            "get": {
                "tags": [
                    "piso-controller"
                ],
                "summary": "getPisoById",
                "operationId": "getPisoByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "numpiso",
                        "in": "path",
                        "description": "numpiso",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PisoDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/roles": {
            "get": {
                "tags": [
                    "role-controller"
                ],
                "summary": "getRoles",
                "operationId": "getRolesUsingGET",
                "produces": [
                    "*/*"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RoleDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/roles/users/{userId}": {
            "get": {
                "tags": [
                    "role-controller"
                ],
                "summary": "geRoleByUserId",
                "operationId": "geRoleByUserIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "userId",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/RoleDto"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/roles/{id}": {
            "get": {
                "tags": [
                    "role-controller"
                ],
                "summary": "geRoleById",
                "operationId": "geRoleByIdUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RoleDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "summary": "getUserById",
                "operationId": "getUserByIdUsingGET_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "user-controller"
                ],
                "summary": "delete",
                "operationId": "deleteUsingDELETE_9",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "CasillaDto": {
            "type": "object",
            "properties": {
                "cod_empresa": {
                    "type": "string"
                },
                "cod_estante": {
                    "type": "string"
                },
                "cod_paquete": {
                    "type": "string"
                },
                "disponibilidad": {
                    "type": "boolean"
                },
                "fecha_final_mant": {
                    "type": "string",
                    "format": "date"
                },
                "fecha_inicio_mant": {
                    "type": "string",
                    "format": "date"
                },
                "idAlmacen": {
                    "type": "string"
                },
                "idCarga": {
                    "type": "string"
                },
                "idCasilla": {
                    "type": "string"
                },
                "idPiso": {
                    "type": "string"
                },
                "mantActivo": {
                    "type": "boolean"
                }
            },
            "title": "CasillaDto"
        },
...
        "EstanteDto": {
            "type": "object",
            "properties": {
                "cod_almacen": {
                    "type": "string"
                },
                "cod_emp": {
                    "type": "string"
                },
                "cod_estante": {
                    "type": "string"
                },
                "fecha_final_mant": {
                    "type": "string",
                    "format": "date"
                },
                "fecha_inicio_mant": {
                    "type": "string",
                    "format": "date"
                },
                "mantActivo": {
                    "type": "boolean"
                },
                "pisos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PisoDto"
                    }
                }
            },
            "title": "EstanteDto"
        },
...
        "PisoDto": {
            "type": "object",
            "properties": {
                "casillas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CasillaDto"
                    }
                },
                "cod_almacen": {
                    "type": "string"
                },
                "cod_emp": {
                    "type": "string"
                },
                "cod_estante": {
                    "type": "string"
                },
                "fecha_final_mant": {
                    "type": "string",
                    "format": "date"
                },
                "fecha_inicio_mant": {
                    "type": "string",
                    "format": "date"
                },
                "mantActivo": {
                    "type": "boolean"
                },
                "numPiso": {
                    "type": "string"
                }
            },
            "title": "PisoDto"
        },
        "RoleDto": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "roleName": {
                    "type": "string"
                }
            },
            "title": "RoleDto"
        },
...
        "UserDto": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identification": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RoleDto"
                    }
                },
                "username": {
                    "type": "string"
                }
            },
            "title": "UserDto"
        }
    }
}